Move tests mistakenly marked as integration_tests
