+++ tests/integration_tests/storage/test_parameter_sample_types.py
@@ -1,10 +1,8 @@
 import os
 import stat
-from contextlib import ExitStack as does_not_raise
 from multiprocessing import Process
-from pathlib import Path
 from textwrap import dedent
-from typing import Optional, Tuple
+from typing import Tuple
 
 import numpy as np
 import pytest
@@ -12,304 +10,15 @@ import xtgeo
 from flaky import flaky
 
 from ert.cli import ENSEMBLE_SMOOTHER_MODE
-from ert.config import ConfigValidationError, ErtConfig
-from ert.enkf_main import create_run_path, ensemble_context, sample_prior
 from ert.libres_facade import LibresFacade
-from ert.storage import EnsembleAccessor, open_storage
+from ert.storage import open_storage
 from tests.integration_tests.run_cli import run_cli
-
-
-def write_file(fname, contents):
-    with open(fname, mode="w", encoding="utf-8") as fout:
-        fout.writelines(contents)
-
-
-def create_runpath(
-    storage,
-    config,
-    active_mask=None,
-    *,
-    ensemble: Optional[EnsembleAccessor] = None,
-    iteration=0,
-    random_seed: Optional[int] = 1234,
-) -> Tuple[ErtConfig, EnsembleAccessor]:
-    active_mask = [True] if active_mask is None else active_mask
-    ert_config = ErtConfig.from_file(config)
-
-    if ensemble is None:
-        experiment_id = storage.create_experiment(
-            ert_config.ensemble_config.parameter_configuration
-        )
-        ensemble = storage.create_ensemble(
-            experiment_id,
-            name="default",
-            ensemble_size=ert_config.model_config.num_realizations,
-        )
-
-    prior = ensemble_context(
-        ensemble,
-        active_mask,
-        iteration,
-        None,
-        "",
-        ert_config.model_config.runpath_format_string,
-        "name",
-    )
-
-    sample_prior(
-        ensemble,
-        [i for i, active in enumerate(active_mask) if active],
-        random_seed=random_seed,
-    )
-    create_run_path(prior, ert_config.substitution_list, ert_config)
-    return ert_config.ensemble_config, ensemble
 
 
 def load_from_forward_model(ert_config, ensemble):
     facade = LibresFacade.from_config_file(ert_config)
     realizations = [True] * facade.get_ensemble_size()
     return facade.load_from_forward_model(ensemble, realizations, 0)
-
-
-@pytest.fixture
-def storage(tmp_path):
-    with open_storage(tmp_path / "storage", mode="w") as storage:
-        yield storage
-
-
-@pytest.mark.usefixtures("set_site_config")
-@pytest.mark.integration_test
-@pytest.mark.parametrize(
-    "config_str, expected, extra_files, expectation",
-    [
-        (
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt\nRANDOM_SEED 1234",
-            "MY_KEYWORD -0.881423",
-            [],
-            does_not_raise(),
-        ),
-        (
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt INIT_FILES:custom_param%d",
-            "MY_KEYWORD 1.31",
-            [("custom_param0", "MY_KEYWORD 1.31")],
-            does_not_raise(),
-        ),
-        (
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt INIT_FILES:custom_param%d",
-            "MY_KEYWORD 1.31",
-            [("custom_param0", "1.31")],
-            does_not_raise(),
-        ),
-        (
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt INIT_FILES:custom_param0",  # noqa
-            "Not expecting a file",
-            [],
-            pytest.raises(
-                ConfigValidationError, match="Loading GEN_KW from files requires %d"
-            ),
-        ),
-    ],
-)
-def test_gen_kw(storage, tmpdir, config_str, expected, extra_files, expectation):
-    with tmpdir.as_cwd():
-        config = dedent(
-            """
-        JOBNAME my_name%d
-        NUM_REALIZATIONS 1
-        """
-        )
-        config += config_str
-        with open("config.ert", mode="w", encoding="utf-8") as fh:
-            fh.writelines(config)
-        with open("template.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD <MY_KEYWORD>")
-        with open("prior.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD NORMAL 0 1")
-        for fname, contents in extra_files:
-            write_file(fname, contents)
-
-        with expectation:
-            create_runpath(storage, "config.ert")
-            assert (
-                Path("simulations/realization-0/iter-0/kw.txt").read_text(
-                    encoding="utf-8"
-                )
-                == expected
-            )
-
-
-@pytest.mark.usefixtures("set_site_config")
-@pytest.mark.integration_test
-@pytest.mark.parametrize(
-    "config_str, expected, extra_files",
-    [
-        pytest.param(
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt",
-            "MY_KEYWORD -0.881423\nNOT KEYWORD <DONT_REPLACE>",
-            [["template.txt", "MY_KEYWORD <MY_KEYWORD>\nNOT KEYWORD <DONT_REPLACE>"]],
-            id="Second magic string that should not be replaced",
-        ),
-        pytest.param(
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt",
-            "MY_KEYWORD -0.881423\n-- if K<=28 then blah blah",
-            [["template.txt", "MY_KEYWORD <MY_KEYWORD>\n-- if K<=28 then blah blah"]],
-            id="Comment in file with <",
-        ),
-        pytest.param(
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt",
-            "MY_KEYWORD -0.881423\nNR_TWO 0.654691",
-            [
-                ["template.txt", "MY_KEYWORD <MY_KEYWORD>\nNR_TWO <NR_TWO>"],
-                ["prior.txt", "MY_KEYWORD NORMAL 0 1\nNR_TWO NORMAL 0 1"],
-            ],
-            id="Two parameters",
-        ),
-    ],
-)
-def test_gen_kw_templating(
-    storage,
-    tmpdir,
-    config_str,
-    expected,
-    extra_files,
-):
-    with tmpdir.as_cwd():
-        config = dedent(
-            """
-        JOBNAME my_name%d
-        NUM_REALIZATIONS 1
-        RANDOM_SEED 1234
-        """
-        )
-        config += config_str
-        with open("config.ert", mode="w", encoding="utf-8") as fh:
-            fh.writelines(config)
-        with open("prior.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD NORMAL 0 1")
-        for fname, contents in extra_files:
-            write_file(fname, contents)
-        create_runpath(storage, "config.ert")
-        assert (
-            Path("simulations/realization-0/iter-0/kw.txt").read_text(encoding="utf-8")
-            == expected
-        )
-
-
-@pytest.mark.usefixtures("set_site_config")
-@pytest.mark.integration_test
-@pytest.mark.parametrize(
-    "relpath",
-    [
-        "somepath/",
-        # This test was added to show current behaviour for Ert.
-        # If absolute paths should be possible to be used like this is up for debate.
-        "/tmp/somepath/",  # ert removes leading '/'
-    ],
-)
-def test_gen_kw_outfile_will_use_paths(tmpdir, storage, relpath: str):
-    with tmpdir.as_cwd():
-        config = dedent(
-            f"""
-        JOBNAME my_name%d
-        NUM_REALIZATIONS 1
-        GEN_KW KW_NAME template.txt {relpath}kw.txt prior.txt
-        """
-        )
-
-        with open("config.ert", mode="w", encoding="utf-8") as fh:
-            fh.writelines(config)
-        with open("template.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD <MY_KEYWORD>")
-        with open("prior.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD NORMAL 0 1")
-        if relpath.startswith("/"):
-            relpath = relpath[1:]
-        create_runpath(storage, "config.ert")
-        assert os.path.exists(f"simulations/realization-0/iter-0/{relpath}kw.txt")
-
-
-@pytest.mark.usefixtures("set_site_config")
-@pytest.mark.integration_test
-@pytest.mark.parametrize(
-    "config_str, expected, extra_files",
-    [
-        (
-            "GEN_KW KW_NAME template.txt kw.txt prior.txt INIT_FILES:custom_param%d",
-            "MY_KEYWORD 1.31\nMY_SECOND_KEYWORD 1.01",
-            [("custom_param0", "MY_SECOND_KEYWORD 1.01\nMY_KEYWORD 1.31")],
-        ),
-    ],
-)
-def test_that_order_of_input_in_user_input_is_abritrary_for_gen_kw_init_files(
-    tmpdir, config_str, expected, extra_files, storage
-):
-    with tmpdir.as_cwd():
-        config = dedent(
-            """
-        JOBNAME my_name%d
-        NUM_REALIZATIONS 1
-        """
-        )
-        config += config_str
-        with open("config.ert", mode="w", encoding="utf-8") as fh:
-            fh.writelines(config)
-        with open("template.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines(
-                "MY_KEYWORD <MY_KEYWORD>\nMY_SECOND_KEYWORD <MY_SECOND_KEYWORD>"
-            )
-        with open("prior.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD NORMAL 0 1\nMY_SECOND_KEYWORD NORMAL 0 1")
-        for fname, contents in extra_files:
-            write_file(fname, contents)
-
-        create_runpath(storage, "config.ert")
-        assert (
-            Path("simulations/realization-0/iter-0/kw.txt").read_text("utf-8")
-            == expected
-        )
-
-
-@pytest.mark.usefixtures("set_site_config")
-@pytest.mark.integration_test
-@pytest.mark.parametrize("load_forward_init", [True, False])
-def test_gen_kw_forward_init(tmpdir, storage, load_forward_init):
-    with tmpdir.as_cwd():
-        config = dedent(
-            """
-        JOBNAME my_name%d
-        NUM_REALIZATIONS 1
-        GEN_KW KW_NAME template.txt kw.txt prior.txt """
-            f"""FORWARD_INIT:{str(load_forward_init)} INIT_FILES:custom_param%d
-        """
-        )
-        with open("config.ert", mode="w", encoding="utf-8") as fh:
-            fh.writelines(config)
-
-        with open("template.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD <MY_KEYWORD>")
-        with open("prior.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD NORMAL 0 1")
-        if not load_forward_init:
-            write_file("custom_param0", "1.31")
-
-        if load_forward_init:
-            with pytest.raises(
-                ConfigValidationError,
-                match=(
-                    "Loading GEN_KW from files created by "
-                    "the forward model is not supported."
-                ),
-            ):
-                create_runpath(storage, "config.ert")
-        else:
-            _, fs = create_runpath(storage, "config.ert")
-            assert Path("simulations/realization-0/iter-0/kw.txt").exists()
-            value = (
-                fs.load_parameters("KW_NAME", 0)
-                .sel(names="MY_KEYWORD")["values"]
-                .values
-            )
-            assert value == 1.31
 
 
 @pytest.mark.usefixtures("set_site_config")
@@ -568,34 +277,3 @@ def run_poly():
         "smoother_update",
         "config.ert",
     )
-
-
-@pytest.mark.usefixtures("set_site_config")
-@pytest.mark.integration_test
-@pytest.mark.parametrize(
-    "config_str, expected",
-    [
-        (
-            "GEN_KW KW_NAME prior.txt\nRANDOM_SEED 1234",
-            -0.881423,
-        ),
-    ],
-)
-def test_gen_kw_optional_template(storage, tmpdir, config_str, expected):
-    with tmpdir.as_cwd():
-        config = dedent(
-            """
-        JOBNAME my_name%d
-        NUM_REALIZATIONS 1
-        """
-        )
-        config += config_str
-        with open("config.ert", mode="w", encoding="utf-8") as fh:
-            fh.writelines(config)
-        with open("prior.txt", mode="w", encoding="utf-8") as fh:
-            fh.writelines("MY_KEYWORD NORMAL 0 1")
-
-        create_runpath(storage, "config.ert")
-        assert list(storage.ensembles)[0].load_parameters("KW_NAME")[
-            "values"
-        ].values.flatten().tolist() == pytest.approx([expected])
